Add Psycho Cut as Gallades signature attack
Flavor-wise it makes a lot of sense for Gallade to get Psycho Cut. It`s one of  It would also give him a more accurate, powerful alternative to zen headbutt. Gallade is near the middle of RU, which is fine, but Mega Gallade is at the bottom of OU. Mega Gallade has no benefits at all from what NEXT has so far, and takes a mega slot, so of all OU pokemon he is the one most needing of something to help him.